Some cleanup and structure refactoring TODO:Figure out why there are two bucket tabs
