+++ src/main/resources/pack.mcmeta
@@ -1,7 +1,7 @@
 {
     "pack": {
         "description": "chem lib resources",
-        "pack_format": 6,
+        "pack_format": 9,
         "_comment": ""
     }
 }
